refactor(retry): move duplicated participant-selection into a shared package
pkg/frost/retry/retry.go was a byte-for-byte copy of pkg/tecdsa/retry/retry.go --
two hand-synchronized copies of the security-critical retry participant-selection
algorithm (EvaluateRetryParticipantsForSigning / ...ForKeyGeneration). A fix
applied to one and not the other would silently make ECDSA and FROST select
different qualified-operator sets for the same seed.

This selection is structurally shared across schemes: it is the scheme-agnostic
"pick the base signing group from the ready members" used for the initial attempt
of every signing (and for DKG). FROST's per-attempt robustness diverges only on
retries, via ROAST (NextAttempt); the initial selection does not -- ROAST is a
transition-from-previous function and cannot produce attempt 0, so the initial
selection is permanently shared.

Move the algorithm (byte-for-byte, verified identical) into a neutral package,
pkg/protocol/retry, and point both callers (pkg/tbtc/dkg_loop.go and
signing_loop_legacy_selector.go -- the only two) at it. Delete pkg/tecdsa/retry
and pkg/frost/retry. Keep the superset test (the tecdsa copy had an extra
triplet-seat-count case the frost copy lacked). No behavior change; a single
source eliminates the drift hazard.
