+++ pkg/tbtc/dkg_loop.go
@@ -11,8 +11,8 @@ import (
 	"github.com/ipfs/go-log/v2"
 	"github.com/keep-network/keep-core/pkg/chain"
 	"github.com/keep-network/keep-core/pkg/protocol/group"
+	"github.com/keep-network/keep-core/pkg/protocol/retry"
 	"github.com/keep-network/keep-core/pkg/tecdsa/dkg"
-	"github.com/keep-network/keep-core/pkg/tecdsa/retry"
 	"golang.org/x/exp/slices"
 )
